move AnswerList to folder and add CSS Module
